Remove LocalCache
While it was now only used for data which was barely changed, it's safer
to delete it to discourage future usage. The problem with local caching
is invalidating the cache when the data changes in other processes.
