+++ packages/server/graphql/queries/helpers/getPublicScoredTemplates.ts
@@ -1,11 +1,14 @@
-import db from '../../../db'
+import {DataLoaderWorker} from '../../graphql'
 import getTemplateScore from '../../../utils/getTemplateScore'
 
 const getPublicScoredTemplates = async (
-  templates: {createdAt: Date; id: string; isStarter?: boolean}[]
+  templates: {createdAt: Date; id: string; isStarter?: boolean}[],
+  dataLoader: DataLoaderWorker
 ) => {
   const sharedTemplateIds = templates.map(({id}) => id)
-  const sharedTemplateEndTimes = await db.readMany('endTimesByTemplateId', sharedTemplateIds)
+  const sharedTemplateEndTimes = await dataLoader
+    .get('endTimesByTemplateId')
+    .load(sharedTemplateIds)
   const scoreByTemplateId = {} as {[templateId: string]: number}
   templates.forEach((template, idx) => {
     const {id: templateId, createdAt, isStarter} = template
